refactor: Entkopplung der app.js in dedizierte Module (Issue #130)
Diese Änderung adressiert Issue #130 zur Verbesserung der Code-Architektur
durch Modularisierung der monolithischen app.js Datei.

Änderungen:
- app.js von IIFE zu ES6-Modul konvertiert (6065 → 5611 Zeilen, -454 Zeilen)
- Neue Manager erstellt:
  * CookieConsentManager.js - Cookie-Consent-Funktionen
  * StorageManager.js - Zentralisiertes localStorage-Management
- Neue Komponenten extrahiert:
  * MermaidDiagram.js - Diagramm-Rendering
  * BrainsPlan.js - "World Domination" Easter Egg
  * RandomRoomButton.js - Zufallsraum-Auswahl mit Tux
  * DolphinMascot.js - Delfin-Maskottchen SVG
- constants.js erweitert mit fehlenden Konstanten:
  * DOOM_CHANNEL_LABEL
  * COOKIE_CONSENT_STORAGE_KEY
  * WHISPER_MODEL_STORAGE_KEY
  * FRANCONIA_INTRO_SEEN_KEY
  * WHISPER_MODELS
  * DEFAULT_WHISPER_MODEL
  * CONSENT_CATEGORIES
- Duplizierten Code entfernt und durch Imports ersetzt
- Single Responsibility Principle wird nun besser eingehalten

Alle bestehenden Funktionen bleiben erhalten. Die Modularisierung
schafft eine Basis für zukünftige Features wie E2EE und WCAG-Konformität.
