+++ index.html
@@ -29,7 +29,7 @@
   <script src="flappy-bird-game.js" defer></script>
   <script src="doom-game.js" defer></script>
   <script src="trivia-ui.js" defer></script>
-  <script src="app.js" defer></script>
+  <script type="module" src="app.js" defer></script>
 </head>
 <body>
   <div id="root"></div>
